Features

+++ src/controller/mainwindow/MainController.java
@@ -4,10 +4,15 @@ import controller.ShowAlert;
 import controller.dbqueries.ArticleQueries;
 import controller.dbqueries.CustomerQueries;
 import controller.dbqueries.ExceptionLog;
+import controller.dbqueries.LoanQueries;
 import javafx.beans.property.Property;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
 import javafx.event.Event;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TextField;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
@@ -42,15 +47,46 @@ public class MainController implements Initializable {
     private ArticleTable articleTable = null;
     private ShowCustomerTable customerTable;
 
+    @FXML
+    private Button deleteArticle;
+
+    @FXML
+    private Button deleteCustomer;
+
     private Property<Staff> loggedInStaff;
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
         try {
             customerTable = new ShowCustomerTable();
+            customerTable.getCustomerProperty().addListener(new ChangeListener<Customer>() {
+                @Override
+                public void changed(ObservableValue<? extends Customer> observableValue, Customer oldValue, Customer newValue) {
+                    /*
+                        if value is null --> hide deleteCustomer
+                        else
+                            show
+                     */
+                    deleteCustomer.setVisible(newValue != null);
+                }
+            });
+            //set visibility of deleteCustomer to false after init
+            deleteCustomer.setVisible(false);
             customerPane.setCenter(customerTable);
 
             articleTable = new ArticleTable();
+            articleTable.getArticleProperty().addListener(new ChangeListener<Article>() {
+                @Override
+                public void changed(ObservableValue<? extends Article> observableValue, Article oldValue, Article newValue) {
+                    /*
+                        if value is null --> hide deleteCustomer
+                        else
+                            show
+                     */
+                    deleteArticle.setVisible(newValue != null);
+                }
+            });
+            deleteArticle.setVisible(false);
             articlePane.setCenter(articleTable);
         } catch (IOException e){
             ExceptionLog.write(e);
@@ -93,6 +129,20 @@ public class MainController implements Initializable {
         }
     }
 
+    public void deleteCustomerMouseClicked(MouseEvent event){
+
+        Customer selectedCustomer = customerTable.getSelectedCustomer();
+        if(LoanQueries.getActiveCustomerLoans(selectedCustomer.getCid()).size() > 0)
+            ShowAlert.showInformation("Der Kunde hat noch nicht alle Artikel zurückgegeben!");
+        else if(ShowAlert.showConfirmation(
+                "Kunde " + selectedCustomer.getNameString() + " wirklich löschen?"
+            ) == ButtonType.YES) {
+
+            CustomerQueries.deleteCustomer(selectedCustomer);
+            customerTable.updateData();
+        }
+    }
+
     @FXML
     void addArticleNameKeyPressed(KeyEvent event) {
         if(event.getCode() == KeyCode.ENTER)
@@ -117,6 +167,21 @@ public class MainController implements Initializable {
         }
     }
 
+    public void deleteArticleMouseClicked(MouseEvent event){
+
+        Article selectedArticle = articleTable.getSelectedArticle();
+        if(!selectedArticle.isAvailable())
+            ShowAlert.showInformation("Artikel ist derzeit verliehen. Löschen nicht möglich!");
+        else if(
+            ShowAlert.showConfirmation(
+                    "Artikel " + selectedArticle.getName() + " wirklich löschen?"
+            ) == ButtonType.YES){
+
+            ArticleQueries.deleteArticle(selectedArticle);
+            articleTable.updateData();
+        }
+    }
+
     public void setLoggedInStaff(Property<Staff> loggedInStaff) {
         this.loggedInStaff = loggedInStaff;
         //set loggedInStaff further down
@@ -130,6 +195,8 @@ public class MainController implements Initializable {
          */
         if(articleTable != null)
             articleTable.updateData();
+        //hide delete button
+        deleteArticle.setVisible(false);
     }
 
     @FXML
@@ -139,24 +206,30 @@ public class MainController implements Initializable {
          */
         if(customerTable != null)
             customerTable.updateData();
+        //hide delete button
+        deleteCustomer.setVisible(false);
     }
     /*
         delete database, initialize empty one
      */
     @FXML
     void deleteData(MouseEvent event) {
-        // delete database
-        Database.delete();
-        // initialize new database
-        Database.init();
+        // ask if they really want to do it
+        if(ShowAlert.showConfirmation("Wirklich alles löschen?") == ButtonType.YES) {
+            // delete database
+            Database.delete();
+            // initialize new database
+            Database.init();
 
-        ShowAlert.showInformation("Alle Daten gelöscht!");
+            ShowAlert.showInformation("Alle Daten gelöscht!");
+        }
     }
     @FXML
     void addData(MouseEvent event) {
-        // add standard data to DB
-        Database.addData();
-
-        ShowAlert.showInformation("Daten hinzugefügt!");
+        if (ShowAlert.showConfirmation("Daten hinzufügen?") == ButtonType.YES) {
+            // add standard data to DB
+            Database.addData();
+            ShowAlert.showInformation("Daten hinzugefügt!");
+        }
     }
 }
